Fix harmless compiler warnings, mostly caused by prior efforts to get static analyzers to run without warnings.
FossilOrigin-Name: cddd3b382a118bb16be3a602ae711fe2c2a7d1c673dd30dd3977b67ee555e666

+++ src/date.c
@@ -1025,7 +1025,7 @@ static void strftimeFunc(
   computeYMD_HMS(&x);
   for(i=j=0; zFmt[i]; i++){
     if( zFmt[i]!='%' ) continue;
-    if( j<i ) sqlite3_str_append(&sRes, zFmt+j, i-j);
+    if( j<i ) sqlite3_str_append(&sRes, zFmt+j, (int)(i-j));
     i++;
     j = i + 1;
     switch( zFmt[i] ){
@@ -1101,7 +1101,7 @@ static void strftimeFunc(
       }
     }
   }
-  if( j<i ) sqlite3_str_append(&sRes, zFmt+j, i-j);
+  if( j<i ) sqlite3_str_append(&sRes, zFmt+j, (int)(i-j));
   sqlite3ResultStrAccum(context, &sRes);
 }
